♻️ Begins to refactor for media
For #4
