+++ composer.json
@@ -1,0 +1,23 @@
+{
+  "name": "shikiryu/shikiryurss",
+  "description": "A RSS reader and writer",
+  "minimum-stability": "stable",
+  "license": "proprietary",
+  "authors": [
+    {
+      "name": "clement",
+      "email": "email@example.com"
+    }
+  ],
+  "autoload": {
+    "psr-4": {
+      "Shikiryu\\SRSS\\": "src/"
+    }
+  },
+  "require-dev": {
+    "phpunit/phpunit": "^9"
+  },
+  "require": {
+    "ext-dom": "*"
+  }
+}
