+++ src/Media/Group.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Shikiryu\SRSS\Media;
+
+class Group
+{
+
+}
