+++ tests/resources/basic.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<rss version="2.0">
+
+    <channel>
+        <title>test Home Page</title>
+        <link>https://www.test.com</link>
+        <description>Free web building tutorials</description>
+        <item>
+            <title>RSS Tutorial</title>
+            <link>https://www.test.com/xml/xml_rss.asp</link>
+            <description>New RSS tutorial on test</description>
+        </item>
+        <item>
+            <title>XML Tutorial</title>
+            <link>https://www.test.com/xml</link>
+            <description>New XML tutorial on test</description>
+        </item>
+    </channel>
+
+</rss>
