+++ tests/resources/media/music-video.xml
@@ -1,0 +1,32 @@
+<rss version="2.0" xmlns:media="http://search.yahoo.com/mrss/"
+     xmlns:dcterms="http://purl.org/dc/terms/">
+
+    <channel>
+        <title>Music Videos 101</title>
+        <link>http://www.foo.com</link>
+        <description>Discussions of great videos</description>
+        <item>
+            <title>The latest video from an artist</title>
+            <link>http://www.foo.com/item1.htm</link>
+            <media:content url="http://www.foo.com/movie.mov" fileSize="12216320" type="video/quicktime"
+                           expression="full">
+                <media:player url="http://www.foo.com/player?id=1111" height="200" width="400"/>
+                <media:hash algo="md5">dfdec888b72151965a34b4b59031290a</media:hash>
+                <media:credit role="producer">producer's name</media:credit>
+                <media:credit role="artist">artist's name</media:credit>
+                <media:category scheme="http://blah.com/scheme">
+                    music/artistname/album/song
+                </media:category>
+                <media:text type="plain">
+                    Oh, say, can you see, by the dawn's early light
+                </media:text>
+                <media:rating>nonadult</media:rating>
+                <dcterms:valid>
+                    start=2002-10-13T09:00+01:00;
+                    end=2002-10-17T17:00+01:00;
+                    scheme=W3C-DTF
+                </dcterms:valid>
+            </media:content>
+        </item>
+    </channel>
+</rss>
